Started dot product.

+++ src/activations.h
@@ -2,8 +2,8 @@
 #ifndef ACTIVATIONS_H
 #define ACTIVATIONS_H
 
-double relu(double x);
-double sigmoid(double x);
-double softmax(double* logits, int num_classes);
+float relu(float x);
+float sigmoid(float x);
+float softmax(float* logits, int num_classes);
 
 #endif // ACTIVATIONS_H
