+++ src/loss.c
@@ -1,7 +1,7 @@
 // loss.c
 #include "loss.h"
 
-double sparse_categorical_crossentropy(int true_label, double* logits, int num_classes) {
+float sparse_categorical_crossentropy(int true_label, float* logits, int num_classes) {
     // Implementation of sparse categorical crossentropy
     // ...
 }
