+++ src/loss.h
@@ -2,7 +2,7 @@
 #ifndef LOSS_H
 #define LOSS_H
 
-double sparse_categorical_crossentropy(int true_label, double* logits, int num_classes);
+float sparse_categorical_crossentropy(int true_label, float* logits, int num_classes);
 // Other loss functions
 
 #endif // LOSS_H
