+++ src/optimizer.h
@@ -2,7 +2,7 @@
 #ifndef OPTIMIZER_H
 #define OPTIMIZER_H
 
-void sgd_optimizer(double* parameters, double* gradients, int num_parameters, double learning_rate);
+void sgd_optimizer(float* parameters, float* gradients, int num_parameters, double learning_rate);
 // Other optimizers
 
 #endif // OPTIMIZER_H
